feat: InfoView 추가
