+++ TarTanning/TarTanning/Presentations/Info/View/InfoView.swift
@@ -1,0 +1,18 @@
+//
+//  InfoView.swift
+//  TarTanning
+//
+//  Created by Jun on 7/28/25.
+//
+
+import SwiftUI
+
+struct InfoView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    InfoView()
+}
